+++ packages/bluesky_cli/lib/src/version.g.dart
@@ -1,2 +1,2 @@
 // This file is generated. Do not manually edit.
-String version = '0.3.3';
+String version = '0.3.5';